Adds ShiftCreator module and moves applicable methods into it
