+++ lib/cypher_master.rb
@@ -1,7 +1,8 @@
 require_relative 'shift_manager'
+require_relative 'shift_creator'
 
 class CypherMaster
-  include ShiftManager
+  include ShiftManager, ShiftCreator
 
   attr_reader :characters
